before handling get_next_line leaks

+++ utils/cleanup/trash_can.c
@@ -6,7 +6,7 @@
 /*   By: moel-hai <moel-hai@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/25 17:20:48 by moel-hai          #+#    #+#             */
-/*   Updated: 2026/01/05 20:35:00 by moel-hai         ###   ########.fr       */
+/*   Updated: 2026/01/06 15:18:22 by moel-hai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,12 +37,13 @@ void	close_fds(void)
 		close (i++);
 }
 
-void	free_everything(t_heap *h, int exit_value)
+void	free_everything(t_heap *h, int exit_value, int _destroy_mlx)
 {
 	if (!h)
 		return ;
 	close_fds();
-	destroy_mlx(configs(0, 0));
+	if (_destroy_mlx)
+		destroy_mlx(configs(0, 0));
 	clear_trash(h);
 	if (exit_value >= 0)
 		exit (exit_value);
